update API and location based on PR feedback
