+++ crates/notion-core/src/distro/node.rs
@@ -4,7 +4,7 @@ use std::fs::{rename, File};
 use std::path::PathBuf;
 use std::string::ToString;
 
-use super::super::ensure_dir_exists;
+use super::super::path::ensure_containing_dir_exists;
 use super::{Distro, Fetched};
 use catalog::NodeCollection;
 use distro::error::DownloadError;
@@ -54,7 +54,7 @@ impl Distro for NodeDistro {
             return NodeDistro::cached(version, File::open(cache_file).unknown()?);
         }
 
-        ensure_dir_exists(&cache_file)?;
+        ensure_containing_dir_exists(&cache_file)?;
         Ok(NodeDistro {
             archive: node_archive::fetch(url, &cache_file)
                 .with_context(DownloadError::for_version(version.to_string()))?,
