+++ crates/notion-core/src/distro/yarn.rs
@@ -4,7 +4,7 @@ use std::fs::{rename, File};
 use std::path::PathBuf;
 use std::string::ToString;
 
-use super::super::ensure_dir_exists;
+use super::super::path::ensure_containing_dir_exists;
 use super::{Distro, Fetched};
 use catalog::YarnCollection;
 use distro::error::DownloadError;
@@ -55,7 +55,7 @@ impl Distro for YarnDistro {
             return YarnDistro::cached(version, File::open(cache_file).unknown()?);
         }
 
-        ensure_dir_exists(&cache_file)?;
+        ensure_containing_dir_exists(&cache_file)?;
         Ok(YarnDistro {
             archive: node_archive::fetch(url, &cache_file)
                 .with_context(DownloadError::for_version(version.to_string()))?,
